describe() and other methods moved from ColumnFile to ActualizeMixin
